extract filtered list logic
