+++ src/components/organisms/CharactersList/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useFilteredCharacters } from './useFilteredCharacters';
